Update IML command output mesages
